requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
bcrypt
pyjwt